Attivati bottoni nuovi

+++ app/src/main/java/com/boymask/testpay/retrofit_boot/PaymentRequest.java
@@ -5,9 +5,9 @@ package com.boymask.testpay.retrofit_boot;
 public class PaymentRequest {
 
     private String productName;
-    private int amount;
+    private long amount;
 
-    public PaymentRequest(String productName, int amount) {
+    public PaymentRequest(String productName, long amount) {
         this.productName = productName;
         this.amount = amount;
     }
